requirements.txt
paddleocr==3.3.0
paddlepaddle==3.2.0
pymupdf
qrdet
numpy==1.25.2
scipy==1.11.3
ultralytics
fastapi
uvicorn[standard]
python-multipart
pillow
beautifulsoup4
opencv-python
huggingface-hub
python-jose[cryptography]
passlib[bcrypt]
pydantic
pydantic-settings
openai
lxml
